django/requirements.txt
Django==2.2
django-cors-headers==2.5.3
django-filter==2.1.0
djangorestframework==3.9.3
Markdown==3.1
pytz==2019.1
sqlparse==0.3.0